+++ src/React.Core/StylesheetHelpers/StyledComponentsFunctions.cs
@@ -3,7 +3,8 @@ using System;
 namespace React
 {
 	/// <summary>
-	/// Render functions for styled components
+	/// Render functions for styled components. https://github.com/styled-components/styled-components
+	/// Requires `styled-components` to be exposed globally as `Styled`
 	/// </summary>
 	public class StyledComponentsFunctions : RenderFunctions
 	{
